+++ oalg-base/src/OAlg/Entity/VectorG.hs
@@ -16,7 +16,11 @@
 --
 -- generalized vectors.
 module OAlg.Entity.VectorG
-  ( VectorG(..)
+  (
+    VectorG(), vectorG
+  , vecGSheaf, vecGpsq
+  , VectorGForm(..)
+
   )
 
   where
@@ -24,6 +28,8 @@ module OAlg.Entity.VectorG
 import Data.List (foldl)
 
 import OAlg.Prelude
+
+import OAlg.Data.Constructable
 
 import OAlg.Structure.Exception
 import OAlg.Structure.Fibred
@@ -35,36 +41,85 @@ import OAlg.Entity.Sequence.PSequence
 --------------------------------------------------------------------------------
 -- VectorG -
 
+-- | generalized vector with elements in a 'Additive' structure @__x__@.
+--
+-- __Property__ Let @v@ be in @t'VectorG' __i x__@ for a 'Additive' structure @__x__@, then holds:
+--
+-- (1) the underlying sheaf of @v@ is 'valid'.
+--
+-- (2) all components of the underlying partial sequence of @v@ are non 'zero'.
 data VectorG i x = VectorG (Root x) (PSequence i x)
 
 deriving instance (Fibred x, Show i) => Show (VectorG i x)
 deriving instance (Fibred x, Eq i) => Eq (VectorG i x)
 deriving instance (Fibred x, Ord x, OrdRoot x, Ord i) => Ord (VectorG i x)
 
+--------------------------------------------------------------------------------
+-- vecGpsq -
+
+-- | the underlying partial sequence.
+vecGpsq :: VectorG i x -> PSequence i x
+vecGpsq (VectorG _ xis) = xis
+
+--------------------------------------------------------------------------------
+-- vecSheaf -
+
+-- | the underlying sheaf.
+vecGSheaf :: VectorG i x -> Sheaf x
+vecGSheaf (VectorG r xis) = Sheaf r (amap1 fst $ psqxs xis) 
+
+--------------------------------------------------------------------------------
+-- Validable -
 
 instance (Additive x, Entity i, Ord i) => Validable (VectorG i x) where
-  valid (VectorG r xs) = Label "VectorG" :<=>:
-    And [ valid xs
-        , foldl (vld r) SValid (amap1 fst $ psqxs xs)
+  valid (VectorG r xis) = Label "VectorG" :<=>:
+    And [ valid r
+        , valid xis
+        , foldl (vld r) SValid (amap1 fst $ psqxs xis)
         ] where
-
-    vld :: Additive x => Root x -> Statement -> x -> Statement
-    vld r s x = And [ s
-                    , Label "Root" :<=>: (root x == r) :?> Params ["r":=show r,"x":=show x]
-                    , Label "non Zero" :<=>: not (isZero x) :?> Params ["x":=show x]
-                    ]
-
+    
+    vld r s x = s && And [ valid x
+                         , Label "root" :<=>: (r == root x) :?> Params ["r":=show r, "x":=show x]
+                         , Label "non zero" :<=>: not (isZero x) :?> Params ["x":=show x]
+                         ]
 
 --------------------------------------------------------------------------------
 -- VectorGForm -
 
-data VectorGForm i x = VectorGForm (Root x) [(x,i)]
+-- | form for constructing vectors.
+data VectorGForm i x = VectorGForm (Root x) (PSequence i x)
+
+deriving instance (Fibred x, Show i) => Show (VectorGForm i x)
+deriving instance (Fibred x, Eq i) => Eq (VectorGForm i x)
+deriving instance (Fibred x, Ord x, OrdRoot x, Ord i) => Ord (VectorGForm i x)
+
+instance (Fibred x, Entity i, Ord i) => Validable (VectorGForm i x) where
+  valid (VectorGForm r xis) = Label "VectorGForm" :<=>:
+    And [ valid r
+        , valid xis
+        , foldl (vld r) SValid (amap1 fst $ psqxs xis)
+        ] where
+
+    vld r s x = s && (Label "root" :<=>: (r == root x) :?> Params ["r":=show r, "x":=show x])
 
 
 --------------------------------------------------------------------------------
 -- vectorG -
 
--- vectorG :: 
+-- | constructing a vector.
+vectorG :: Additive x => VectorGForm i x -> VectorG i x
+vectorG (VectorGForm r xis) = VectorG r $ psqFilter (not . isZero) xis
+
+--------------------------------------------------------------------------------
+-- Constructable -
+
+instance Exposable (VectorG i x) where
+  type Form (VectorG i x) = VectorGForm i x
+  form (VectorG r xis) = VectorGForm r xis
+
+instance Additive x => Constructable (VectorG i x) where
+  make = vectorG
+
 --------------------------------------------------------------------------------
 -- Additve - Abelian - Vectorial -