+++ oalg-base/src/OAlg/Hom/Fibred.hs
@@ -23,7 +23,8 @@ module OAlg.Hom.Fibred
   (
     -- * Fibred
     HomFibred, FunctorialFibred
-
+  , HomFib, HomFibEmpty
+  
     -- * Duality
   , DualisableFibred
   , toDualStk, toDualRt
@@ -74,6 +75,19 @@ instance (HomFibred h, Disjunctive2 h)  => HomFibred (Variant2 v h)
 instance HomFibred h => HomFibred (Inv2 h)
 
 instance (Transformable s Fbr, HomFibred h) => HomFibred (Sub s h)
+
+--------------------------------------------------------------------------------
+-- HomFib -
+
+-- | fibred homomorphism on structure @__s__@.
+type HomFib h = Path h
+
+--------------------------------------------------------------------------------
+-- HomFibEmpty -
+
+-- | fibred homomorphism on structure @__s__@.
+type HomFibEmpty s = HomFib (HomEmpty s)
+
 
 --------------------------------------------------------------------------------
 -- DualisableFibred -